Add files via upload
Employee Management System is an application that enables users to create and store Employees Records. Thus there is a need of a system which can handle the data of such a large number of Employees in a organization. This project simplifies the task of maintain records because of its user friendly nature.